chore: fix auth tests
