+++ router-tests/block_operations_test.go
@@ -1,4 +1,4 @@
-package integration_test
+package integration
 
 import (
 	"bytes"
